pyproject.toml
[project]
name = "motion-canvas-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bs4>=0.0.2",
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "gitpython>=3.1.44",
    "jinja2>=3.1.6",
    "langchain>=0.3.26",
    "langchain-anthropic>=0.3.17",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.68",
    "langchain-google-genai>=2.1.7",
    "langchain-openai>=0.3.27",
    "langgraph>=0.5.2",
    "openai>=1.95.1",
    "panel>=1.7.4",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "streamlit>=1.46.1",
]
